move state, BottomBarLayout to internal state package
